+++ src/ldap_codec/encoder.rs
